+++ internal/tchart/gauge_int_test.go
@@ -8,7 +8,7 @@ import (
 
 func TestComputeDeltas(t *testing.T) {
 	uu := map[string]struct {
-		d1, d2 int
+		d1, d2 int64
 		e      delta
 	}{
 		"same": {
